The Keto Restaurant App
The app that I decided to do for my capstone is called the Keto app. This is an app for all those keto users out there that have trouble decided on where they want to eat out. 
It is easy to find recipes for keto but when it comes to finding keto friendly dishes at restaurants it' s different. If my users are like me it takes awhile to figure out what I want to eat and how many carbs and calories are in my overall food order.